Add battle command modules
