+++ commands/player/cmd_battle_attack.py
@@ -1,3 +1,5 @@
+"""Command for queuing a move in battle."""
+
 from __future__ import annotations
 
 import re
@@ -11,27 +13,14 @@ except Exception:  # pragma: no cover
     EnhancedEvMenu = None  # type: ignore
 
 from utils.battle_display import render_move_gui
+from .cmd_battle_utils import NOT_IN_BATTLE_MSG, _get_participant
 
-NOT_IN_BATTLE_MSG = "You are not currently in battle."
-
-try:
+try:  # pragma: no cover - battle engine may not be available in tests
     from pokemon.battle import Action, ActionType, BattleMove
-    if Action is None or ActionType is None or BattleMove is None:
+    if Action is None or ActionType is None or BattleMove is None:  # type: ignore[truthy-bool]
         raise ImportError
 except Exception:  # pragma: no cover - fallback if engine isn't loaded
     from pokemon.battle.engine import Action, ActionType, BattleMove
-
-
-def _get_participant(inst, caller):
-    """Return battle participant for caller or fallback to first."""
-    if inst and getattr(inst, "battle", None):
-        for part in getattr(inst.battle, "participants", []):
-            if getattr(part, "player", None) is caller:
-                return part
-        if getattr(inst.battle, "participants", []):
-            return inst.battle.participants[0]
-    return None
-
 
 
 class CmdBattleAttack(Command):
@@ -61,7 +50,13 @@ class CmdBattleAttack(Command):
         if not getattr(self.caller.db, "battle_control", False):
             self.caller.msg("|rWe aren't waiting for you to command right now.")
             return
-        from pokemon.battle.battleinstance import BattleSession
+        try:  # pragma: no cover - battle session may be absent in tests
+            from pokemon.battle.battleinstance import BattleSession
+        except Exception:  # pragma: no cover
+            class BattleSession:  # type: ignore[override]
+                @staticmethod
+                def ensure_for_player(caller):
+                    return getattr(caller.ndb, "battle_instance", None)
 
         inst = BattleSession.ensure_for_player(self.caller)
         if not inst or not inst.battle:
@@ -244,7 +239,7 @@ class CmdBattleAttack(Command):
                     self.caller,
                     battle_move_menu,
                     startnode="start",
-                    cmd_on_exit=None,              # don't auto-look after menu ends
+                    cmd_on_exit=None,  # don't auto-look after menu ends
                     start_kwargs=dict(
                         slots=slots,
                         pp_overrides=pp_overrides,
@@ -253,7 +248,7 @@ class CmdBattleAttack(Command):
                     ),
                     numbered_options=False,
                     show_options=False,
-                    show_footer=True,              # shown on input nodes; hidden on terminal nodes by formatter
+                    show_footer=True,  # shown on input nodes; hidden on terminal nodes by formatter
                     menu_title="Move Select",
                     footer_prompt="A–D or name",
                     invalid_message="Invalid. Type A–D, exact name, or 'quit'.",
@@ -263,197 +258,4 @@ class CmdBattleAttack(Command):
             return
 
         _process_move(move_name)
-
-
-class CmdBattleSwitch(Command):
-    """Switch your active Pokémon in battle.
-
-    Usage:
-      +battle/switch <slot>
-    """
-
-    key = "+battle/switch"
-    aliases = ["+switch", "+Switch", "+battleswitch"]
-    locks = "cmd:all()"
-    help_category = "Pokemon/Battle"
-
-    def func(self):
-        if not getattr(self.caller.db, "battle_control", False):
-            self.caller.msg("|rWe aren't waiting for you to command right now.")
-            return
-        slot = self.args.strip()
-        from pokemon.battle.battleinstance import BattleSession
-
-        inst = BattleSession.ensure_for_player(self.caller)
-        if not inst or not inst.battle:
-            self.caller.msg(NOT_IN_BATTLE_MSG)
-            return
-        participant = _get_participant(inst, self.caller)
-
-        if not slot:
-            lines = []
-            for idx, poke in enumerate(participant.pokemons, 1):
-                status = " (fainted)" if getattr(poke, "hp", 0) <= 0 else ""
-                active = " (active)" if poke in participant.active else ""
-                lines.append(f"{idx}. {poke.name}{active}{status}")
-            lines.append("0. Cancel")
-
-            def _callback(caller, prompt, result):
-                choice = result.strip().lower()
-                if choice in {"0", "cancel", "quit", "exit", "abort", ".abort"}:
-                    caller.msg("Switch cancelled.")
-                    return False
-                try:
-                    idx = int(choice) - 1
-                    poke = participant.pokemons[idx]
-                except (ValueError, IndexError):
-                    caller.msg("Invalid Pokémon slot.")
-                    return True
-                if poke in participant.active:
-                    caller.msg(f"{poke.name} is already active.")
-                    return True
-                if getattr(poke, "hp", 0) <= 0:
-                    caller.msg(f"{poke.name} has fainted and cannot battle.")
-                    return True
-                action = Action(participant, ActionType.SWITCH, priority=6)
-                action.target = poke
-                participant.pending_action = action
-                caller.msg(f"You prepare to switch to {poke.name}.")
-                if hasattr(inst, "queue_switch"):
-                    try:
-                        inst.queue_switch(idx + 1, caller=self.caller)
-                    except Exception:
-                        pass
-                elif hasattr(inst, "maybe_run_turn"):
-                    try:
-                        inst.maybe_run_turn()
-                    except Exception:
-                        pass
-                return False
-
-            prompt = "Choose a Pokémon to switch to or 0 to cancel:\n" + "\n".join(lines)
-            get_input(self.caller, prompt, _callback)
-            return
-
-        if slot.lower() in {"0", "cancel", "quit", "exit", "abort", ".abort"}:
-            self.caller.msg("Switch cancelled.")
-            return
-        try:
-            index = int(slot) - 1
-            pokemon = participant.pokemons[index]
-        except (ValueError, IndexError):
-            self.caller.msg("Invalid Pokémon slot.")
-            return
-        if pokemon in participant.active:
-            self.caller.msg(f"{pokemon.name} is already active.")
-            return
-        if getattr(pokemon, "hp", 0) <= 0:
-            self.caller.msg(f"{pokemon.name} has fainted and cannot battle.")
-            return
-        action = Action(participant, ActionType.SWITCH, priority=6)
-        action.target = pokemon
-        participant.pending_action = action
-        self.caller.msg(f"You prepare to switch to {pokemon.name}.")
-        if hasattr(inst, "queue_switch"):
-            try:
-                inst.queue_switch(index + 1, caller=self.caller)
-            except Exception:
-                pass
-        elif hasattr(inst, "maybe_run_turn"):
-            try:
-                inst.maybe_run_turn()
-            except Exception:
-                pass
-
-
-class CmdBattleItem(Command):
-    """Use an item during battle.
-
-    Usage:
-      +battle/item <item>
-    """
-
-    key = "+battle/item"
-    aliases = ["+item", "+Item", "+battleitem"]
-    locks = "cmd:all()"
-    help_category = "Pokemon/Battle"
-
-    def func(self):
-        if not getattr(self.caller.db, "battle_control", False):
-            self.caller.msg("|rWe aren't waiting for you to command right now.")
-            return
-        item_name = self.args.strip()
-        if not item_name:
-            self.caller.msg("Usage: +battle/item <item>")
-            return
-        if not self.caller.has_item(item_name):
-            self.caller.msg(f"You do not have any {item_name}.")
-            return
-        from pokemon.battle.battleinstance import BattleSession
-
-        inst = BattleSession.ensure_for_player(self.caller)
-        if not inst or not inst.battle:
-            self.caller.msg(NOT_IN_BATTLE_MSG)
-            return
-        participant = _get_participant(inst, self.caller)
-        target = inst.battle.opponent_of(participant)
-        action = Action(
-            participant,
-            ActionType.ITEM,
-            target,
-            item=item_name,
-            priority=6,
-        )
-        participant.pending_action = action
-        if hasattr(self.caller, "trainer"):
-            self.caller.trainer.remove_item(item_name)
-        self.caller.msg(f"You prepare to use {item_name}.")
-        if hasattr(inst, "queue_item"):
-            try:
-                inst.queue_item(item_name, caller=self.caller)
-            except Exception:
-                pass
-        elif hasattr(inst, "maybe_run_turn"):
-            try:
-                inst.maybe_run_turn()
-            except Exception:
-                pass
-
-
-class CmdBattleFlee(Command):
-    """Attempt to flee from battle.
-
-    Usage:
-      +battle/flee
-    """
-
-    key = "+battle/flee"
-    aliases = ["+flee", "+Flee", "+battleflee"]
-    locks = "cmd:all()"
-    help_category = "Pokemon/Battle"
-
-    def func(self):
-        if not getattr(self.caller.db, "battle_control", False):
-            self.caller.msg("|rWe aren't waiting for you to command right now.")
-            return
-        from pokemon.battle.battleinstance import BattleSession
-
-        inst = BattleSession.ensure_for_player(self.caller)
-        if not inst or not inst.battle:
-            self.caller.msg(NOT_IN_BATTLE_MSG)
-            return
-        participant = _get_participant(inst, self.caller)
-        action = Action(participant, ActionType.RUN, priority=9)
-        participant.pending_action = action
-        self.caller.msg("You attempt to flee!")
-        if hasattr(inst, "queue_run"):
-            try:
-                inst.queue_run(caller=self.caller)
-            except Exception:
-                pass
-        elif hasattr(inst, "maybe_run_turn"):
-            try:
-                inst.maybe_run_turn()
-            except Exception:
-                pass
