+++ tests/test_battle_item_command.py
@@ -10,8 +10,8 @@ sys.path.insert(0, ROOT)
 
 
 def load_cmd_module():
-    path = os.path.join(ROOT, "commands", "player", "cmd_battle.py")
-    spec = importlib.util.spec_from_file_location("commands.player.cmd_battle", path)
+    path = os.path.join(ROOT, "commands", "player", "cmd_battle_item.py")
+    spec = importlib.util.spec_from_file_location("commands.player.cmd_battle_item", path)
     mod = importlib.util.module_from_spec(spec)
     sys.modules[spec.name] = mod
     spec.loader.exec_module(mod)
